AI_Server/requirements.txt
# Video Captioning Dependencies
# Python 3.10+ recommended

# Deep Learning Frameworks
tensorflow==2.15.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# CLIP Model
open-clip-torch>=2.20.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0,<2.0.0

# Utilities
tqdm>=4.65.0

# NLP Evaluation
nltk>=3.8.0

# Azure Service Bus for Queue
azure-servicebus>=7.11.0

# HTTP requests for callback
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
